Fix missing template file error and restructure command line api
