+++ iris/conclude.py
@@ -1,0 +1,2 @@
+def start(project, debug):
+    print("conclude mode is not yet implemented.")
